Fix context manager exception handling
When using a connection object as a context manager (via a with
statement), and the context raises an exception, the __exit__ method
should simply return a value other than True in order for the exception
to be raised. The current handling manually re-raises the exception, and
also creates a new exception object from the raised one in a way that
does not always work (e.g. if there are more than one arguments to the
exception's constructor). To resolve this we allow the context manager
to raise the exception.

Fixes: #601